refactored error handlers
