+++ .eslintrc.json
@@ -15,7 +15,7 @@
   "rules": {
     "no-unused-vars": "error",
     "no-console": "error",
-    // "no-undef": "error",
+    "no-undef": "error",
     "no-unused-expressions": "error",
     "no-unreachable": "error"
     // "consistent-type-definitions": ["error", "type"]
@@ -23,6 +23,10 @@
 
   "env": {
     "browser": true,
-    "es2021": true
+    "es2021": true,
+    "node": true
+  },
+  "globals": {
+    "process": "readonly"
   }
 }
