+++ src/app.ts
@@ -1,7 +1,8 @@
 import cors from 'cors'
 import express, { Application, Request, Response } from 'express'
 import globalErrorHandler from './app/middlewares/globalErrorHandler'
-import userRouter from './app/modules/users/users.route'
+import { UserRoutes } from './app/modules/users/user.route'
+
 const app: Application = express()
 
 app.use(cors())
@@ -11,7 +12,7 @@ app.use(express.json())
 app.use(express.urlencoded({ extended: false }))
 
 // application routes
-app.use('/api/v1/users/', userRouter)
+app.use('/api/v1/users/', UserRoutes)
 
 // testing route
 app.get('/', (req: Request, res: Response) => {
