+++ src/app/modules/users/user.controller.ts
@@ -1,0 +1,20 @@
+import { RequestHandler } from 'express'
+import { UserService } from './user.services'
+
+const createUser: RequestHandler = async (req, res, next) => {
+  try {
+    const { user } = req.body
+    const result = await UserService.createUser(user)
+    res.status(200).json({
+      success: true,
+      message: 'user created successfully!',
+      data: result,
+    })
+  } catch (err) {
+    next(err)
+  }
+}
+
+export const UserController = {
+  createUser,
+}
