+++ src/app/modules/users/user.interface.ts
@@ -1,0 +1,10 @@
+import { Model } from 'mongoose'
+
+export type IUser = {
+  id: string
+  role: string
+  password: string
+}
+
+// type UserModel = Model<IUser, object, IUserMethods>
+export type UserModel = Model<IUser, Record<string, unknown>>
