--- src/app/modules/users/users.controller.ts
@@ -1,20 +1,0 @@
-import { NextFunction, Request, Response } from 'express'
-import usersServices from './users.services'
-
-const createUser = async (req: Request, res: Response, next: NextFunction) => {
-  try {
-    const { user } = req.body
-    const result = await usersServices.createUser(user)
-    res.status(200).json({
-      success: true,
-      message: 'user created successfully!',
-      data: result,
-    })
-  } catch (err) {
-    next(err)
-  }
-}
-
-export default {
-  createUser,
-}
