--- src/app/modules/users/users.interface.ts
@@ -1,5 +1,0 @@
-export type IUser = {
-  id: string
-  role: string
-  password: string
-}
