--- src/app/modules/users/users.route.ts
@@ -1,7 +1,0 @@
-import express from 'express'
-import usersController from './users.controller'
-const router = express.Router()
-
-router.post('/create-user', usersController.createUser)
-
-export default router
